admin profile

+++ POSS.Services/CartServices/Implementation/CartService.cs
@@ -44,7 +44,6 @@ namespace POSS.Services.CartServices.Implementation
                     }
 
                     //Find the product on the cart
-                    double productPrice = 0;
 
                     var product = dbContext.Products.Where(x => x.Id == model.Id).SingleOrDefault();
 
@@ -104,13 +103,17 @@ namespace POSS.Services.CartServices.Implementation
             {
                 using (dbContext)
                 {
-                    var product = dbContext.Products.Find(Id);
-                    var cart = dbContext.Carts.Where(x => x.ProductId == Id && x.UserCartId == Username).SingleOrDefault();
-
-                    if (cart != null)
+                    if(Username != null)
                     {
-                        dbContext.Carts.Remove(cart);
-                        dbContext.SaveChanges();
+                        var product = dbContext.Products.Find(Id);
+                        var cart = dbContext.Carts.Where(x => x.ProductId == Id && x.UserCartId == Username).SingleOrDefault();
+
+                        if (cart != null)
+                        {
+                            dbContext.Carts.Remove(cart);
+                            dbContext.SaveChanges();
+                        }
+
                     }
                 }
                 return $"Item removed from cart !";
